modified authentication functionality in a way that one cannot add new dream items unless they're logged in

+++ frontend/src/components/DreamDetails.tsx
@@ -1,6 +1,7 @@
 import { Dream } from "../types";
 import { useDreamsContext } from "../hooks/useDreamsContext";
 import formatDistanceToNow from "date-fns/formatDistanceToNow";
+import { useAuthContext } from "../hooks/useAuthContext";
 
 interface DreamDetailProps {
     dream: Dream;
@@ -9,11 +10,18 @@ interface DreamDetailProps {
 const DreamDetails = (props: DreamDetailProps) => {
     const { title, description, createdAt, _id } = props.dream;
     const { dispatch } = useDreamsContext();
+    const { user } = useAuthContext();
     // console.log(props);
 
     const handleDelete = async () => {
+        if (!user) {
+            return
+        }
         const response = await fetch('http://localhost:4000/api/dreams/' + _id, {
-            method: 'DELETE'
+            method: 'DELETE',
+            headers: {
+                'Authorization': `Bearer ${user.token}`
+            }
         })
         const json = await response.json();
